split files and separate duration validation from reschedule
