+++ qiskit/transpiler/passes/scheduling/__init__.py
@@ -16,5 +16,11 @@ from .alap import ALAPSchedule
 from .asap import ASAPSchedule
 from .time_unit_conversion import TimeUnitConversion
 from .dynamical_decoupling import DynamicalDecoupling
-from .instruction_alignment import AlignMeasures, ValidatePulseGates, ConstrainedReschedule
 from .padding import PadDelay
+from .alignments import InstructionDurationCheck, ValidatePulseGates, ConstrainedReschedule
+
+# For backward compability
+from . import alignments as instruction_alignments
+
+# TODO Deprecated pass. Will be removed after deprecation period.
+from .alignments import AlignMeasures
